@@ -6,13 +6,14 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
+import java.util.List;
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.joining;
 
 @NoArgsConstructor
 public class CamelCaseDisplayNameGenerator extends DisplayNameGenerator.Standard {
-    private static final String NESTED_ICON = "\u21AA";
+    private static final String NESTED_ICON = "↪";
     private static final String BOOKS_ICON = "\uD83D\uDCDA";
     private static final String EXPERT_ICON = "\uD83D\uDC74";
 
@@ -32,14 +33,14 @@ public class CamelCaseDisplayNameGenerator extends DisplayNameGenerator.Standard
     }
 
     @Override
-    public String generateDisplayNameForNestedClass(Class<?> nestedClass) {
-        var className = super.generateDisplayNameForNestedClass(nestedClass);
+    public String generateDisplayNameForNestedClass(List<Class<?>> enclosingInstanceTypes, Class<?> nestedClass) {
+        var className = super.generateDisplayNameForNestedClass(enclosingInstanceTypes, nestedClass);
         return NESTED_ICON + " " + splitByCamelCase(className)
                 .collect(joining(" "));
     }
 
     @Override
-    public String generateDisplayNameForMethod(Class<?> testClass, Method testMethod) {
+    public String generateDisplayNameForMethod(List<Class<?>> enclosingInstanceTypes, Class<?> testClass, Method testMethod) {
         return splitByCamelCase(testMethod.getName())
                 .map(String::toLowerCase)
                 .collect(joining(" "));
